ai-agent/requirements.txt
# AI Agentic Pull Request System Dependencies
# Core dependencies for the AI agent functionality

# Configuration handling
PyYAML>=6.0

# Optional dependencies for enhanced functionality
# Uncomment based on your needs:

# GitHub API integration
# PyGithub>=1.58.0
# requests>=2.28.0

# AI/ML libraries (if using local models)
# openai>=1.0.0
# anthropic>=0.3.0

# Code analysis
# pylint>=2.17.0
# bandit>=1.7.0  # Security linting

# Testing
# pytest>=7.0.0
# pytest-cov>=4.0.0

# Development tools
# black>=23.0.0  # Code formatting
# isort>=5.12.0  # Import sorting